+++ Runtime/Scripts/IPathfinding.cs
@@ -6,7 +6,7 @@
  * Date Created: February 14, 2023
  * 
  * Additional Comments:
- *      File Line Length: 120
+ *      File Line Length: 140
  */
 
 using System.Collections;
@@ -26,17 +26,30 @@ namespace Kokowolo.Pathfinding
         /************************************************************/
         #region Functions
 
-        public List<Node> GetNeighborsFromNode(Node node);
+        // NOTE: this may change based on pathfinding configuration, i.e. a node's unit using its GetNeighbors() vs what's below
+        public List<Node> GetNeighborsFromNode(Node node)
+        {
+            return node.GetNeighbors();
+        }
 
-        public bool IsValidMoveBetweenNodes(Node start, Node end);
+        public bool IsValidMoveBetweenNodes(Node start, Node end)
+        {
+            return start.HasNeighbor(end);
+        }
 
+        // NOTE: this should get the distance/approximate between these two nodes
         public int GetHeuristicCostBetweenNodes(Node start, Node end);
 
+        // NOTE: this should get the actual distance between these two nodes
         public int GetMoveCostBetweenNodes(Node start, Node end);
 
-        public bool IsPathOutsideMovementRange(NodePath path);
+        // NOTE: this method is a hacky way of checking if the current path is too long; i.e. path.Distance > maxDistance;
+        public bool IsPathOutsideMovementRange(NodePath path)
+        {
+            return false;
+        }
 
-        // TODO: add burst or jobs to project
+        // TODO: add burst or jobs to package
         // internal void OnSearchComplete(PathfindingNodePath path);
         
         #endregion