+++ Runtime/Scripts/NodeDebugObject.cs
@@ -6,7 +6,7 @@
  * Date Created: August 22, 2022
  * 
  * Additional Comments:
- *		File Line Length: 120
+ *		File Line Length: 140
  */
 
 using System.Collections;