+++ Runtime/Scripts/NodePath.cs
@@ -6,7 +6,7 @@
  * Date Created: February 13, 2023
  * 
  * Additional Comments:
- *      File Line Length: 120
+ *      File Line Length: 140
  */
 
 using System.Collections;